refactor(domain): decompose ports.py into ports/ package
Split monolithic ports.py (781 LOC) into organized package:
- ports/__init__.py: re-exports for backward compatibility
- ports/storage.py: StoragePort (~242 LOC)
- ports/data_source.py: DataSourcePort, FilterableDataSourcePort
- ports/locking.py: LockPort
- ports/checkpoint.py: CheckpointPort
- ports/quarantine.py: QuarantinePort
- ports/observability.py: TracingPort, MetricsPort, LoggerPort
- ports/validation.py: GoldValidatorPort
- ports/filtering.py: InputFilterPort

All existing imports via `from bioetl.domain.ports import X`
continue to work unchanged (backward compatible).

Updated architecture tests to work with new package structure.
